slight updates and refactoring
The players now stop 30 pixels away from the camera. I cleaned up the
code a little bit. V2.2 is the current version of the game. It is
currently set up to incorporate lighting from the 4 players and ambient
light. Wands are missing. It is also set up to incorporate collisions.