Spells freeze opponents, collisions kindof works
Collisions only works 90% of the time. This causes the opponents to get
stuck in walls so I allow then to still walk through walls. I added a
feature where the spells can hit the opponents causing them to freeze
and turn black. Not sure if they should freeze for a random period of
time or not. For now they freeze indefinitely.